Update REAMDE.md to add The Herd Locker
Based on the hundreds of thousands of #robotics tweets already processed, noise is a little over 95% of the conversation. The platform filters and ranks the most popular shared content in realtime. Robotics' zeitgeist, you might say. It's been running for over a year, monitoring half a billion tweets a day, tracking dozens of topics. It will always be free to use. No ads. No BS.